[not verified] WIP version of Mapkit support
