+++ projects/plugins/jetpack/extensions/blocks/map/sources/index.js
@@ -1,0 +1,3 @@
+import MapBoxBlock from './mapbox';
+import MapkitBlock from './mapkit';
+export { MapBoxBlock, MapkitBlock };
