+++ projects/plugins/jetpack/extensions/blocks/map/style.scss
@@ -8,6 +8,11 @@
 		min-height: 400px;
 		text-align: left;
 	}
+
+  .wp-block-jetpack-map__mb-container {
+	height: 400px;
+  }
+
 	.mapboxgl-popup {
 		h3 {
 			font-size: 1.3125em;
